+++ processor/src/chiplets/hasher/mod.rs
@@ -387,7 +387,7 @@ impl Hasher {
         lookups: &mut Vec<HasherLookup>,
     ) -> Word {
         assert!(!path.is_empty(), "path is empty");
-        assert!(index >> path.len() == 0, "invalid index for the path");
+        assert!(index >> path.len().min(63) == 0, "invalid index for the path");
         let mut root = value;
         let mut depth = path.len() - 1;